+++ src/components/index.js
@@ -6,7 +6,7 @@ export * from './inits/initDeleteCompleted.js'
 // export * from './inits/initDelete.js'
 
 export * from './ui/updateTasksOrder.js'
-// export * from './ui/addNewTodo.js'
+export * from './ui/addNewTodo.js'
 // export * from './ui/updateTask.js'
 // export * from './ui/renderData.js'
-export * from './ui/loadData.js'
+// export * from './ui/loadData.js'
