+++ src/components/ui/addNewTodo.js
@@ -1,0 +1,29 @@
+import { showError } from '../../utils/helpers.js'
+import { addTodo } from '../../API/index.js'
+// import { loadData } from '../index.js'
+import { loadData } from '../../app.js'
+
+export async function addNewTodo(taskInput) {
+  const newTodoText = taskInput.value.trim()
+
+  if (!newTodoText) {
+    alert('Введите текст задачи!')
+    return
+  }
+
+  const newTodo = {
+    text: newTodoText,
+    createdAt: Date.now(),
+    completed: false
+  }
+
+  try {
+    await addTodo(newTodo)
+    console.log('Задача добавлена')
+    taskInput.value = ''
+    await loadData()
+  } catch (error) {
+    console.error(error.message)
+    showError('Не удалось добавить задачу')
+  }
+}
